Create separate directory for `TimeRangeFilter` components.
