+++ graylog2-web-interface/src/components/configurations/QuickAccessTimeRangeForm.tsx
@@ -26,7 +26,7 @@ import { Icon, SortableList } from 'components/common';
 import type { TimeRange } from 'views/logic/queries/Query';
 import { useStore } from 'stores/connect';
 import { SearchConfigStore } from 'views/stores/SearchConfigStore';
-import TimeRangeFilter from 'views/components/searchbar/TimeRangeFilter';
+import TimeRangeFilter from 'views/components/searchbar/time-range-filter';
 import TimeRangeInputSettingsContext from 'views/components/contexts/TimeRangeInputSettingsContext';
 import generateId from 'logic/generateId';
