+++ graylog2-web-interface/src/views/components/WidgetQueryControls.tsx
@@ -62,7 +62,7 @@ import useAppDispatch from 'stores/useAppDispatch';
 import useHandlerContext from 'views/components/useHandlerContext';
 
 import TimeRangeOverrideInfo from './searchbar/WidgetTimeRangeOverride';
-import TimeRangeFilter from './searchbar/TimeRangeFilter';
+import TimeRangeFilter from './searchbar/time-range-filter';
 import StreamsFilter from './searchbar/StreamsFilter';
 import SearchButton from './searchbar/SearchButton';
 import QueryInput from './searchbar/queryinput/AsyncQueryInput';
@@ -191,12 +191,12 @@ const WidgetQueryControls = ({ availableStreams }: Props) => {
               <ValidateOnParameterChange parameters={parameters} />
               <TimeRangeRow>
                 {!hasTimeRangeOverride && (
-                <TimeRangeFilter disabled={hasTimeRangeOverride}
-                                 limitDuration={limitDuration}
-                                 onChange={(nextTimeRange) => setFieldValue('timerange', nextTimeRange)}
-                                 value={values?.timerange}
-                                 hasErrorOnMount={!!errors.timerange}
-                                 position="right" />
+                  <TimeRangeFilter disabled={hasTimeRangeOverride}
+                                   limitDuration={limitDuration}
+                                   onChange={(nextTimeRange) => setFieldValue('timerange', nextTimeRange)}
+                                   value={values?.timerange}
+                                   hasErrorOnMount={!!errors.timerange}
+                                   position="right" />
                 )}
                 {hasTimeRangeOverride && (
                   <TimeRangeOverrideInfo value={globalOverride?.timerange}
